Use CGI.espace and CGI.unescape instead of URI for actual compatibility with the RFC spec and add options to skip various steps in encoding procedure of HTTPRedirect binding.
Background :

URI.encode wasn't properly encoding `+` characters (left them unescaped)
and this caused problems when trying to decode the string.
For example on a rails app (possibly because of params decoding?) `+`
was converted to white space.

CGI encode actually fixes this since it encodes it properly to `%2B`.

For further discussion on URI encodings in ruby you can check these :

* http://stackoverflow.com/questions/2824126/whats-the-difference-between-uri-escape-and-cgi-escape
* http://bugs.ruby-lang.org/issues/2542
* http://addressable.rubyforge.org/

And more general :

* http://tools.ietf.org/html/rfc3986
* http://www.blooberry.com/indexdot/html/topics/urlencoding.htm#rfc1738